requirements.txt
pyTelegramBotAPI==4.3.1
Flask==2.0.2
psycopg2-binary==2.9.3